+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 *[M]: major change; [m]: minor change*
 
+* [m] Allow customisation of the command delimiter
+* [m] Allow scalars in loops
+* [m] Allow access to iteration index in loops, e.g. `$company._idx` (where `company` is the name of the loop variable)
+
 ## 1.0.1, 1.0.2 (Dec. 15, 2016)
 
-* Add docs.
+* [M] Add docs
 
 ## 1.0.0 (Dec. 15, 2016)
 
-* First public release.
+* [M] First public release